[#67] refactor: components 에 있는 컴포넌트 app 안으로 이동
