+++ src/app/debateList/components/Comment.tsx
